Add delete and random button
Also remove the slang list

+++ SlangDetailPane.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class SlangDetailPane extends JPanel {
 	JTextField slangField;
 	ArrayList<JTextField> meaningFields;
-	JButton saveButton, editButton, addButton;
+	JButton saveButton, deleteButton, editButton, addButton;
 	
 	SlangDictionary dictionary;
 
@@ -48,9 +48,19 @@ public class SlangDetailPane extends JPanel {
 			public void actionPerformed(ActionEvent event) {
 				changeEditMode(false);
 				try {
-					var slang = slangField.getText();
-					dictionary.deleteSlang(editedSlang);
-					dictionary.addSlang(slang, SlangDetailPane.this.getMeaningsFromFields());
+					var slang = slangField.getText().toUpperCase();
+					boolean updateOk = true;
+					if (dictionary.getMeanings(slang) != null) {
+						int res = JOptionPane.showConfirmDialog(SlangDetailPane.this, "This slang has already existed\nDo you want to update it?");
+						updateOk = res == JOptionPane.OK_OPTION;
+					}
+
+					if (updateOk) {
+						if (editedSlang != null && editedSlang.length() > 0)
+							dictionary.deleteSlang(editedSlang);
+						dictionary.addSlang(slang, SlangDetailPane.this.getMeaningsFromFields());
+						JOptionPane.showMessageDialog(SlangDetailPane.this, "Saved!");
+					}
 				}
 				catch (Exception e) {
 					e.printStackTrace();
@@ -59,6 +69,38 @@ public class SlangDetailPane extends JPanel {
 		});
 
 		buttonPane.add(saveButton);
+
+		deleteButton = new JButton("Delete");
+		deleteButton.setEnabled(false);
+		deleteButton.addActionListener(new ActionListener() {
+			@Override
+			public void actionPerformed(ActionEvent event) {
+				changeEditMode(false);
+				try {
+					var slang = slangField.getText().toUpperCase();
+					boolean updateOk = true;
+					if (dictionary.getMeanings(slang) != null) {
+						int res = JOptionPane.showConfirmDialog(SlangDetailPane.this, "Do you want to delete this slang?");
+						updateOk = res == JOptionPane.OK_OPTION;
+					}
+					else 
+						JOptionPane.showMessageDialog(SlangDetailPane.this, "This slang doesn't exist");
+
+					if (updateOk) {
+						if (slang != null && slang.length() > 0)
+							dictionary.deleteSlang(slang);
+						clearTextFields();
+						JOptionPane.showMessageDialog(SlangDetailPane.this, "Deleted!");
+					}
+					deleteButton.setEnabled(false);
+				}
+				catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
+		});
+
+		buttonPane.add(deleteButton);
 		
 		editButton = new JButton("Edit");
 		editButton.addActionListener(new ActionListener() {
@@ -72,14 +114,38 @@ public class SlangDetailPane extends JPanel {
 		buttonPane.add(editButton);
 		
 		addButton = new JButton("New slang");
+		addButton.addActionListener(new ActionListener() {
+			@Override
+			public void actionPerformed(ActionEvent event) {
+				clearTextFields();
+				changeEditMode(true);
+			}
+		});
 		
 		buttonPane.add(addButton);
+
+		var randomButton = new JButton("Random slang");
+		randomButton.addActionListener(new ActionListener() {
+			@Override
+			public void actionPerformed(ActionEvent event) {
+				var slangs = dict.getRandomSlangs(1);
+				setSlang(slangs.iterator().next());
+			}
+		});
+		
+		buttonPane.add(randomButton);
 
 		add(buttonPane, BorderLayout.SOUTH);
 	}
 
 	public SlangDetailPane(SlangDictionary dict) {
 		this(dict, 3);
+	}
+
+	void clearTextFields() {
+		slangField.setText("");
+		for (var field: meaningFields) 
+			field.setText("");
 	}
 
 	void changeEditMode(boolean edit) {
@@ -101,12 +167,13 @@ public class SlangDetailPane extends JPanel {
 
 	public void setSlang(String slang) {
 		slangField.setText(slang);
+		deleteButton.setEnabled(true);
 		
 		var meanings = dictionary.getMeanings(slang);
 		int lim1 = meaningFields.size(), lim2 = meanings.size();
 		
 		if (lim2 > lim1)
-			System.err.println("Warning: Slang " + slang + " has " + lim2 + " meanigs");
+			System.err.println("Warning: Slang " + slang + " has " + lim2 + " meanings");
 
 		for (int i = 0; i < lim1; i++) {
 			String text = "";
